GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Update task.dart to remove enums
 (refs/gitbutler/Update-task.dart-to-remove-enums)
   branch head: 78667600a97e29bea4a49b5301c04b91540ba1cc
 - Virtual branch (refs/gitbutler/Virtual-branch)
   branch head: 67738be470ea9cbcb272adaf2948133062cc1692
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch